+++ paddle_billing/Resources/DiscountGroups/Operations/ListDiscountGroups.py
@@ -1,34 +1,22 @@
-from paddle_billing.EnumStringify import enum_stringify
 from paddle_billing.HasParameters import HasParameters
-
-from paddle_billing.Entities.Shared import Status
 
 from paddle_billing.Exceptions.SdkExceptions.InvalidArgumentException import InvalidArgumentException
 
 from paddle_billing.Resources.Shared.Operations import Pager
 
 
-class ListDiscounts(HasParameters):
+class ListDiscountGroups(HasParameters):
     def __init__(
         self,
         pager: Pager | None = None,
         ids: list[str] | None = None,
-        statuses: list[Status] | None = None,
-        codes: list[str] | None = None,
-        discount_group_ids: list[str] | None = None,
     ):
         self.pager = pager
         self.ids = ids if ids is not None else []
-        self.statuses = statuses if statuses is not None else []
-        self.codes = codes if codes is not None else []
-        self.discount_group_ids = discount_group_ids if discount_group_ids is not None else []
 
         # Validation
         for field_name, field_value, field_type in [
             ("ids", self.ids, str),
-            ("statuses", self.statuses, Status),
-            ("codes", self.codes, str),
-            ("discount_group_ids", self.discount_group_ids, str),
         ]:
             invalid_items = [item for item in field_value if not isinstance(item, field_type)]
             if invalid_items:
@@ -42,11 +30,5 @@ class ListDiscounts(HasParameters):
             parameters.update(self.pager.get_parameters())
         if self.ids:
             parameters["id"] = ",".join(self.ids)
-        if self.statuses:
-            parameters["status"] = ",".join(map(enum_stringify, self.statuses))
-        if self.codes:
-            parameters["code"] = ",".join(self.codes)
-        if self.discount_group_ids:
-            parameters["discount_group_id"] = ",".join(self.discount_group_ids)
 
         return parameters